backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
fastapi==0.133.1
frozendict==2.4.7
h11==0.16.0
idna==3.11
joblib==1.5.3
multitasking==0.0.12
numpy==2.4.2
pandas==3.0.1
peewee==4.0.0
platformdirs==4.9.2
protobuf==7.34.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
psycopg[binary]==3.2.12
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
xgboost==3.2.0
yfinance==1.2.0
slowapi==0.1.9
slowapi==0.1.9
psycopg2-binary==2.9.9
alembic==1.14.1
